article complete

+++ src/main/java/_thBackEnd/LectureCode/controller/ArticleController.java
@@ -18,6 +18,9 @@ public class ArticleController {
 
     @PostMapping("/article/add")
     public ArticleDTO.ResponseArticle createArticle(@RequestHeader("Authorization") String token, @RequestBody ArticleDTO.addArticleReq request){
+        if (!jwtUtility.validateToken(token)) {
+            return null;
+        }
         String userId = jwtUtility.getClaimsFromToken(token).getSubject();
         Article article = articleService.addArticle(userId, request.getTitle(), request.getContent());
         return new ArticleDTO.ResponseArticle(article);
@@ -25,12 +28,18 @@ public class ArticleController {
 
     @PutMapping("/article/update")
     public ArticleDTO.ResponseArticle updateArticle(@RequestHeader("Authorization") String token, @RequestBody ArticleDTO.ArticleReq request){
+        if (!jwtUtility.validateToken(token)) {
+            return null;
+        }
         Article article = articleService.updateArticle(request.getArticleId(), request.getTitle(), request.getContent(), token);
         return new ArticleDTO.ResponseArticle(article);
     }
 
     @DeleteMapping("/article/{articleId}")
     public void deleteArticle(@RequestHeader("Authorization") String token, @PathVariable("articleId") Long articleId){
+        if (!jwtUtility.validateToken(token)) {
+            return ;
+        }
         articleService.deleteArticle(articleId, token);
     }
